Change to size_t

+++ litert/vendors/qualcomm/dispatch/litert_dispatch_invocation_context.cc
@@ -109,7 +109,7 @@ LiteRtDispatchInvocationContextT::LiteRtDispatchInvocationContextT(
   dump_buffer_.resize(total_bytes);
   size_t byte_offset = 0;
   for (auto it = mid; it != outputs_.end(); ++it) {
-    uint32_t num_bytes = it->GetBytes();
+    size_t num_bytes = it->GetBytes();
     it->SetClientBuf(dump_buffer_.data() + byte_offset, num_bytes);
     byte_offset += num_bytes;
   }
